ingestion/snowflake/requirements.txt
acryl-datahub[snowflake]>=0.10.0
snowflake-connector-python>=3.0.0
sqlalchemy>=1.4.0
requests>=2.27.0
schedule>=1.1.0
slackclient>=2.9.0
pyyaml>=6.0